Fix trophies alignment in board
Extracted Trophies for better dependencies
