+++ src/components/TeamInfo/TeamInfo.scss
@@ -134,20 +134,6 @@
     font-size: 0.75rem;
   }
 
-  .trophies {
-    display: flex;
-    flex-direction: row;
-
-    .trophy {
-      width: 2.75rem;
-
-      font-size: 0.75rem;
-      text-align: left;
-      text-wrap: nowrap;
-      white-space: nowrap;
-    }
-  }
-
   .match-icon {
     margin: 0 0.25rem;
   }
