+++ src/components/TeamInfo/TeamInfo.tsx
@@ -11,8 +11,8 @@ import {
   getTeamStats,
   isTeamInCourse,
 } from '../../model/model-helpers';
-import { getTrophies } from '../CompetitionBoard/CompetitionBoard';
 import MatchSheetLink from '../MatchSheetLink/MatchSheetLink';
+import Trophies from '../Trophies/Trophies';
 
 import './TeamInfo.scss';
 
@@ -60,7 +60,7 @@ const renderTeam = ({ competition, team, day, displayRanking }: TeamInfoProps) =
             <div>
               {team.department.num_dep} - {team.department.region_name}
             </div>
-            <div>{getTrophies(competition, team)}</div>
+            <Trophies competition={competition} team={team} />
             <div className="small-text">
               <div>
                 ranking: {ranking ?? '-'} / {competition.days[day].teams.length} <small>{delta}</small> | points:{' '}
@@ -152,7 +152,7 @@ const renderMatchs = (
                 </a>
               )}
             </div>
-            <div>{getTrophies(competition, teamB)}</div>
+            <Trophies competition={competition} team={teamB} />
           </div>
         </div>
       ),
