+++ tsconfig.json
@@ -23,7 +23,7 @@
     /* Absolute Imports */
     "baseUrl": "./",
     "paths": {
-      "@/*": ["*", "src/*"]
+      "@/*": ["src/*"]
     }
   },
   "include": ["src"],
